Share TreeHeadings/TreeContainer between download tree and archive views
Moves these styled-components out of WorkDetails.Tree.tsx into the
shared work.styles.tsx so the upcoming archive collection contents
table can reuse them too. No behaviour change to the existing
download table.
